tidy up __init__.py
